Refresh ticket class selector
When entering TOP0119A and then changing performances to TOP0202 (or otherwise browsing various performances) all potential ticket classes from both performances will populate as selectable options. This does not appear to be a problem for these specific performances, but for any performance that has assigned seating.

This affects the ticket class selector screen, not the ticket class totals themselves.

- Jeremy Wechsler: ticket-selector is now emptied before being rebuilt by populate_ticket_selector()